libdnf-cli: Move public headers to include/libdnf-cli
