+++ include/libdnf-cli/output/groupinfo.hpp
