+++ include/libdnf-cli/output/grouplist.hpp
