+++ include/libdnf-cli/output/key_value_table.hpp
