+++ include/libdnf-cli/output/repo_info.hpp
