+++ include/libdnf-cli/output/transaction_table.hpp
